+++ Gruntfile.js
@@ -26,7 +26,7 @@ module.exports = function (grunt) {
 				dest: 'dist/polyfill.object-fit.css'
 			},
 			js: {
-				src: ['src/polyfill.getMatchedCSSRules.js', 'src/polyfill.rAF.js', 'src/polyfill.object-fit.core.js'],
+				src: ['src/polyfill.getMatchedCSSRules.js', 'node_modules/raf.js/raf.js', 'src/polyfill.object-fit.core.js'],
 				dest: 'dist/polyfill.object-fit.js'
 			}
 		},
@@ -50,7 +50,7 @@ module.exports = function (grunt) {
 					'Gruntfile.js',
 					'src/polyfill.object-fit.css',
 					'src/polyfill.getMatchedCSSRules.js',
-					'src/polyfill.rAF.js',
+					'node_modules/raf.js/raf.js',
 					'src/polyfill.object-fit.core.js'
 				],
 				tasks: ['jshint', 'concat', 'uglify']