+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### HEAD
 
-* Use [external `element.closest()` polyfill](https://github.com/jonathantneal/closest) (#30), small performance win
+* Use [external `element.closest()` polyfill](https://github.com/jonathantneal/closest) via npm (#30), small performance win
 * Use [external rAF polyfill](https://github.com/ngryman/raf.js) via npm (#30)
 
 ### 0.4.1 (February, 24th, 2015)